added .ui file generated GUI, still some autoUIC issues
