--- src/view/gui/MoveControls.h
@@ -1,43 +1,0 @@
-#ifndef SWAG_SCANNER_MOVECONTROLS_H
-#define SWAG_SCANNER_MOVECONTROLS_H
-
-#include <QWidget>
-
-class QVBoxLayout;
-
-class QFormLayout;
-
-class QLineEdit;
-
-class QPushButton;
-
-class MoveFormsPayload;
-
-class MoveControls : public QWidget {
-Q_OBJECT
-public:
-    explicit MoveControls(QWidget *parent = 0);
-
-signals:
-
-    void move_button_pressed(const MoveFormsPayload &vars);
-
-    void set_home_button_pressed();
-
-private slots:
-
-    void send_move_button_pressed();
-
-    void send_set_home_button_pressed();
-
-private:
-    QVBoxLayout *v_layout;
-    QFormLayout *form_layout;
-    QLineEdit *to_edit;
-    QLineEdit *by_edit;
-    QPushButton *move_button;
-    QPushButton *set_home_button;
-
-};
-
-#endif //SWAG_SCANNER_MOVECONTROLS_H
